fix(refUpdates): Adds defensive check to 'unobserve' method
On updates (re-rendering) there are cases where the ref callback gets called twice: first
to nullify the DOM node and again when setting the new DOM node. ReactDOM.createPortal()
for example causes this when it's re-rendered, causing an exception when we call
unobserve() on a null node. Future updates should change the way we handle refs to a more
solid solution.