fix(reobserve): Prevent unnecessary reobserve() calls when re-rendering
In some cases it's possible that the target node changed and there's a need to reobserve it.
When that happens a flag was stale and subsequent re-renders cause reobserve() to be called always.
This fix sets the flag value back to false on the second re-render so that this behavior gets resetted.